Fixed potential memory leak, using VAOs for render models, misc. changes to render settings, re-raised GLSL version
We were attempting to use sizeof() to determine the sizes of arrays to delete using OGL functions, but the arrays were dynamically allocated and sizeof() was only giving the size of the pointers. Instead of storing a size, each array was limited to being a single value. This has been done before and similarly undone. It has yet to set in stone whether models with multiple objects should be represented with multiple render models or put in he same render model. Older GLSL versions appear to become deprecated with new OpenGL versions. For that reason, the GLSL version has been re-upped from commit db6406c. From here on, we will target the newest OGL deatures. If compatibility is an issue, we can port things backward later. In other words, we're deferring that problem to later. Settings changed include sample object positions, movements, and the projection matrix. This is to be able to see that we have depth and that all those transformatons and things work. We also replaced the test texture. The old one did not appear very well at certain angles.