Work on physics system simulations
Instead of maintaining a single world and adding/removing bodies from it, we're maintaining a number of worlds. One for each Scene, to be exact. Each scene is mapped to a world in the physics system. This relationship is stored as a property of the system and not the scene to simplify world instantiation and keep Bullet code out of Scene. We have a problem, though. Rigid bodies from Actors' collision models must be added to their respective scene's world. This means either passing Bullet types all the way up the compositional hierarchy, or breaking the hierarchy altogether. We either need to find a way around this, or accept Bullet as an inherent part of our application.