Large number of changes revolving around physics.
The key thing that's changed in this commit from the last is that we now have a working physics system. However, that's come at the cost of no longer having our traditional Actor movement work. At the moment, the translation function applies an impulse to the Actor. This is just to display a bit of physics behaviour, it is not to be kept.

Of note otherwise is that the physics system has been moved from a singular component of the game to a per-scene component and given a more specific name that reflects its purpose - PhysicsSimulator. I have also given up on keeping Bullet code away from the application proper. Bullet is now an integral part of the code, for better or worse. As a part of that, PhysicsModel has been removed; it no longer served any useful purpose in abstracting physics away from the rest of the data model.

This is a large commit, so there are likely to be many pieces that require more thorough review.
All in all, however, I believe the changes provided are just about the best way for now to get physics simulation and collision detection into the system.