Fixed copy-safety issue.
The problem was that any copy would leave what were effectively dangling pointers on deletion. It was neither easy nor desirable to duplicate the lost resources, so instead we keep a count on the number of instances sharing the resources and delete them only when the last one is destroyed.