Changed public-facing matrix types to GLM types, instead of pointers to basic type arrays
Passing around pointers to arrays is clearly a horrible idea. It's blatantly unsafe. Ownership is ambiguous, size is ambiguous, properties such as matrix order are ambiguous. The reason we were doing it in the first place was to keep library types out of public interfaces in case we even needed or wanted to change them. But that's plainly stupid to do this way and I am not aware of any compromise that takes care of all the issues involved other than duplicating the GLM types—and then why not just use GLM?