Added material support and 3D texture-loading
The cpp-side of things needs to be updated to reflect the new lighting
model. Specular is also yet to be tested under the new mod
el.

To summarize the new model: Each object *must* have two maps:
color/diffuse and shine/specular. The color map is your standard
texture. The specular map gives the specular color as rgb and the shine
as the alpha channel.

The shader must still be modified to take lighting properties into
account among the other tasks written in as TODOs.