Restructure actor construction, remove SimulatedPhysicsActor, enable PMDL loading
RenderModels are now loaded as part of the Scene, not by each individual
Actor. Actors are constructed by paassing a reference-counted pointer.
Thus the model only needs to be constructed once, and it will be deleted
when all the actors using it are deleted.

The changes generated by modifying Actor have created some
incompatibilites with SimulatedPhysicsActor. Since the physics aspects
of the engine need some rewriting anyway, I've gone ahead and completely
removed the existing implementation.

RenderModels are now created using our internal PMDL format.
Unfortunately, the issues that was previously assumed to be the fault of
tinyobjloader persists. The fault, therefore, lies either with the
export to the intermediate format from Blender or the way our code is
handling UVs. More testing must be done.

Work on this commit also lead to a number of changes to the PMDL
library. The library is currently included in the source directory. The
specification and PDCONV will be updated to reflect these changes. The
version will not be incremented since I'm 100% sure nobody else is using
the format or library at the moment and it's a small bother to do the
change. The only real difference is that the version field has been
extended by three bytes to prevent padding of the Header structure,
which should make other implementations easier.