Make RenderModel dictionary static member of Actor
Doing this keep Actor construction within the Actor and allows us to
follow the instrinsic, recursive method of scene graphing that we were
aiming for before. Instead of requiring the Scene to parse an Actor's
description in order to build the model for it, this commit allows the
Actor to build its own model.