Keep model pointers on c++ side
Prolog files had no need for the memory addresses of the input and
spatial models we were passing, so all the calls we were making to do so
were extraneous.

Input actions are currently the most time-consuming part of the code and
this change resulting in about a 50% reduction in that regard. This is
both a good and a bad thing. First, it's good that this commit made a
50% improvement. Second, it's bad that the way we're going about input
is making it the slowest piece.