Picking works with limitations
First, we have to clear to black. Otherwise, the picking buffer will also be cleared to the background colour and we'll inevitably end up with a false selection. This can be avoided by not giving any object an ID of that colour--we have to do that anyway--but it's easier to avoid an ID of 0 that some arbitrary value.

The second limitation is our encoding scheme. Something was wrong about
our earlier integer-only scheme, so I've applied a standard RGB scheme
instead. It works, but it requires some worrisome floating point
arithmetic and bit manipulation. I haven't gotten to that bit
manipulation yet, so we're stuck using the R value only at the moment.