Modify wrapper access functions
We are not overloading operator->() as said a couple cimmit messages ago. This is open for reconsideration. Geometry offers access to multiple data members so it doesn't make sense to treat it like a pointer. It may be more sensible for ShaderProgram, but do we treat it as a (smart) pointer or a reference?