Implement texture loading
This is amost a straight-up copy from the deleted RenderModel class. We did not switch to libktx for two reasons:
1) Getting the library is a PITA and requires a special build process
2) gli handles both DDS and KTX

The globjects::Texture is constructed from a classic OpenGL resource handle because globjects::Texture is currently missing any compressedSubImage2D function. There are only compressedImage2D and subImage2D. This applies to other dimensions as well.

As with all the current exception handling, specific exceptions deriving from the standard exfeption classes need to be implemented.
Resource loading also needs significant correcting. readFIle() is... not very good.