Do not retain shader sources
Apparently it's fine. I was going by what I read in globjects' `Shader.cpp`, but I guess I overlooked something. Maybe it's fine for fine for `File` but not for a source string? Maybe the detail's in one of the many implementation files?
Who knows? It's not a problem for the time being, so I don't care; I don't need to understand the internals so long as it works according to the API.